backend/requirements.txt
Flask
flask-cors
flask-pymongo
flask-mail
gunicorn
PyJWT
cloudinary
python-dotenv
requests
